+++ fluent/backends/request/client.js
@@ -1,7 +1,11 @@
 'use strict'
 
-const request = require('request')
+const { convertKubeconfig } = require('./config')
+const deprecate = require('depd')('kubernetes-client')
+const JSONStream = require('json-stream')
+const pump = require('pump')
 const qs = require('qs')
+const request = require('request')
 const urljoin = require('url-join')
 const WebSocket = require('ws')
 
@@ -13,7 +17,7 @@ const WebSocket = require('ws')
  */
 function refreshAuth (type, config) {
   return new Promise((resolve, reject) => {
-    const provider = require(`../auth-providers/${type}.js`)
+    const provider = require(`../../lib/auth-providers/${type}.js`)
     provider.refresh(config)
       .then(result => {
         const auth = {
@@ -98,26 +102,36 @@ class Request {
    */
   constructor (options) {
     this.requestOptions = options.request || {}
-    this.requestOptions.qsStringifyOptions = { indices: false }
-    this.requestOptions.baseUrl = options.url
-    this.requestOptions.ca = options.ca
-    this.requestOptions.cert = options.cert
-    this.requestOptions.key = options.key
-    if ('insecureSkipTlsVerify' in options) {
-      this.requestOptions.strictSSL = !options.insecureSkipTlsVerify
+
+    let convertedOptions
+    if (!options.kubeconfig) {
+      deprecate('Request() without a .kubeconfig option, see ' +
+                'https://github.com/godaddy/kubernetes-client/blob/master/merging-with-kubernetes.md')
+      convertedOptions = options
+    } else {
+      convertedOptions = convertKubeconfig(options.kubeconfig)
     }
-    if ('timeout' in options) {
-      this.requestOptions.timeout = options.timeout
+
+    this.requestOptions.qsStringifyOptions = { indices: false }
+    this.requestOptions.baseUrl = convertedOptions.url
+    this.requestOptions.ca = convertedOptions.ca
+    this.requestOptions.cert = convertedOptions.cert
+    this.requestOptions.key = convertedOptions.key
+    if ('insecureSkipTlsVerify' in convertedOptions) {
+      this.requestOptions.strictSSL = !convertedOptions.insecureSkipTlsVerify
+    }
+    if ('timeout' in convertedOptions) {
+      this.requestOptions.timeout = convertedOptions.timeout
     }
 
     this.authProvider = {
       type: null
     }
-    if (options.auth) {
-      this.requestOptions.auth = options.auth
-      if (options.auth.provider) {
-        this.requestOptions.auth = options.auth.request
-        this.authProvider = options.auth.provider
+    if (convertedOptions.auth) {
+      this.requestOptions.auth = convertedOptions.auth
+      if (convertedOptions.auth.provider) {
+        this.requestOptions.auth = convertedOptions.auth.request
+        this.authProvider = convertedOptions.auth.provider
       }
     }
   }
@@ -147,6 +161,17 @@ class Request {
 
       return cb(null, { statusCode: res.statusCode, body: body })
     })
+  }
+
+  async getLogByteStream (options) {
+    return this.http(Object.assign({ stream: true }, options))
+  }
+
+  async getWatchObjectStream (options) {
+    const jsonStream = new JSONStream()
+    const stream = this.http(Object.assign({ stream: true }, options))
+    pump(stream, jsonStream)
+    return jsonStream
   }
 
   /**