+++ fluent/lib/swagger-client.js
@@ -24,12 +24,14 @@
  */
 
 const Component = require('swagger-fluent').Component
+const deprecate = require('depd')('kubernetes-client')
 const fs = require('fs')
+const KubeConfig = require('./config')
 const path = require('path')
 const zlib = require('zlib')
 
 const getAliases = require('./alias')
-const Request = require('./backends/request')
+const Request = require('../backends/request')
 
 class Root extends Component {
   _getSpec (pathname) {
@@ -57,79 +59,151 @@ class Root extends Component {
       })
   }
 
+  _getLogByteStream (options) {
+    return this.backend.getLogByteStream(Object.assign({
+      method: 'GET',
+      pathItemObject: this.pathItemObject,
+      pathname: this.getPath(),
+      pathnameParameters: this.getPathnameParameters()
+    }, options))
+  }
+
+  async _getWatchObjectStream (options) {
+    return this.backend.getWatchObjectStream(Object.assign({
+      method: 'GET',
+      pathItemObject: this.pathItemObject,
+      pathname: this.getPath(),
+      pathnameParameters: this.getPathnameParameters()
+    }, options))
+  }
+
+  _addEndpoint (endpoint) {
+    const component = super._addEndpoint(endpoint)
+    if (!component) return component
+
+    //
+    // Deprecate stream API.
+    //
+    if (endpoint.pathItem.get) {
+      const action = endpoint.pathItem.get['x-kubernetes-action']
+      if (action === 'watch' || action === 'watchlist') {
+        component.getStream = deprecate.function(
+          component.getStream,
+          '.getStream use .getObjectStream, see ' +
+            'https://github.com/godaddy/kubernetes-client/blob/master/merging-with-kubernetes.md')
+        component.getObjectStream = component._getWatchObjectStream
+      } else if (endpoint.name === '/api/v1/namespaces/{namespace}/pods/{name}/log') {
+        component.getStream = deprecate.function(
+          component.getStream,
+          '.getStream use .getByteStream, see ' +
+            'https://github.com/godaddy/kubernetes-client/blob/master/merging-with-kubernetes.md')
+        component.getByteStream = component._getLogByteStream
+      } else {
+        component.getStream = deprecate.function(
+          component.getStream,
+          '.getStream see https://github.com/godaddy/kubernetes-client/blob/master/merging-with-kubernetes.md')
+      }
+    }
+  }
+
   addCustomResourceDefinition (manifest) {
     const group = manifest.spec.group
-    const version = manifest.spec.version
     const name = manifest.spec.names.plural
-    const spec = { paths: {} }
     const namespace = manifest.spec.scope === 'Cluster' ? '' : '/namespaces/{namespace}'
 
-    //
-    // Make just enough of Swagger spec to generate some useful endpoints.
-    //
-    const templatePath = `/apis/${group}/${version}${namespace}/${name}/{name}`
-    spec.paths[templatePath] = ['delete', 'get', 'patch', 'put'].reduce((acc, method) => {
-      acc[method] = { operationId: `${method}Template${name}` }
-      return acc
-    }, {})
-
-    const resourcePath = `/apis/${group}/${version}${namespace}/${name}`
-    spec.paths[resourcePath] = ['get', 'post'].reduce((acc, method) => {
-      acc[method] = { operationId: `${method}${name}` }
-      return acc
-    }, {})
-    //
-    // Namespaced CRDs get a cluster-level GET endpoint.
-    // Similar to GET /api/v1/pods.
-    //
-    if (manifest.spec.scope === 'Namespaced') {
-      const clusterPath = `/apis/${group}/${version}/${name}`
-      spec.paths[clusterPath] = {
-        get: {
-          operationId: `getCluster${name}`
-        }
-      }
-    }
-
-    const watchPaths = {
-      watchCluster: `/apis/${group}/${version}/watch/${name}`,
-      watchNamespace: `/apis/${group}/${version}${namespace}/watch/${name}`,
-      watchResource: `/apis/${group}/${version}${namespace}/watch/${name}/{name}`
-    }
-    Object.keys(watchPaths).forEach(operationId => {
-      const watchPath = watchPaths[operationId]
-      spec.paths[watchPath] = {
-        get: {
-          operationId: `operationId${name}`
-        }
-      }
-    })
-
-    // Add status endpoint - see https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.11/#customresourcesubresourcestatus-v1beta1-apiextensions-k8s-io
-    if (manifest.spec.subresources && manifest.spec.subresources.status) {
-      const statusPath = `/apis/${group}/${version}${namespace}/${name}/{name}/status`
-      spec.paths[statusPath] = ['get', 'put'].reduce((acc, method) => {
+    const addSpec = (version) => {
+      const spec = { paths: {} }
+      //
+      // Make just enough of Swagger spec to generate some useful endpoints.
+      //
+      const templatePath = `/apis/${group}/${version}${namespace}/${name}/{name}`
+      spec.paths[templatePath] = ['delete', 'get', 'patch', 'put'].reduce((acc, method) => {
         acc[method] = { operationId: `${method}Template${name}` }
         return acc
       }, {})
-    }
 
-    // Add scale endpoints - see https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.11/#customresourcesubresourcescale-v1beta1-apiextensions-k8s-io
-    if (manifest.spec.subresources && manifest.spec.subresources.scale) {
-      const statusPath = `/apis/${group}/${version}${namespace}/${name}/{name}/scale`
-      spec.paths[statusPath] = ['get', 'put'].reduce((acc, method) => {
-        acc[method] = { operationId: `${method}Template${name}` }
+      const resourcePath = `/apis/${group}/${version}${namespace}/${name}`
+      spec.paths[resourcePath] = ['get', 'post'].reduce((acc, method) => {
+        acc[method] = { operationId: `${method}${name}` }
         return acc
       }, {})
+      //
+      // Namespaced CRDs get a cluster-level GET endpoint.
+      // Similar to GET /api/v1/pods.
+      //
+      if (manifest.spec.scope === 'Namespaced') {
+        const clusterPath = `/apis/${group}/${version}/${name}`
+        spec.paths[clusterPath] = {
+          get: {
+            operationId: `getCluster${name}`
+          }
+        }
+      }
+
+      const watchPaths = {
+        watchCluster: `/apis/${group}/${version}/watch/${name}`,
+        watchNamespace: `/apis/${group}/${version}/watch${namespace}/${name}`,
+        watchResource: `/apis/${group}/${version}/watch${namespace}/${name}/{name}`
+      }
+      Object.keys(watchPaths).forEach(operationId => {
+        const watchPath = watchPaths[operationId]
+        spec.paths[watchPath] = {
+          get: {
+            operationId: `operationId${name}`
+          }
+        }
+      })
+
+      // Add status endpoint - see https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.11/#customresourcesubresourcestatus-v1beta1-apiextensions-k8s-io
+      if (manifest.spec.subresources && manifest.spec.subresources.status) {
+        const statusPath = `/apis/${group}/${version}${namespace}/${name}/{name}/status`
+        spec.paths[statusPath] = ['get', 'put'].reduce((acc, method) => {
+          acc[method] = { operationId: `${method}Template${name}` }
+          return acc
+        }, {})
+      }
+
+      // Add scale endpoints - see https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.11/#customresourcesubresourcescale-v1beta1-apiextensions-k8s-io
+      if (manifest.spec.subresources && manifest.spec.subresources.scale) {
+        const statusPath = `/apis/${group}/${version}${namespace}/${name}/{name}/scale`
+        spec.paths[statusPath] = ['get', 'put'].reduce((acc, method) => {
+          acc[method] = { operationId: `${method}Template${name}` }
+          return acc
+        }, {})
+      }
+
+      this._addSpec(spec)
     }
 
-    this._addSpec(spec)
+    if (manifest.spec.version) {
+      addSpec(manifest.spec.version)
+    } else {
+      const versions = manifest.spec.versions || []
+      versions.forEach(version => addSpec(version.name))
+    }
   }
 }
 
 class Client {
   constructor (options) {
-    const backend = options.backend || new Request(options.config)
+    options = options || {}
+
+    if (options.config) {
+      deprecate('Client({ config }), see ' +
+                'https://github.com/godaddy/kubernetes-client/blob/master/merging-with-kubernetes.md')
+    }
+
+    let backend = options.backend
+    if (!backend) {
+      if (options.config) {
+        backend = new Request(options.config)
+      } else {
+        const kubeconfig = new KubeConfig()
+        kubeconfig.loadFromDefault()
+        backend = new Request({ kubeconfig })
+      }
+    }
+
     let spec = options.spec
     if (!spec && options.version) {
       const swaggerPath = path.join(